+++ libre_pvz_animation/CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Cargo's Semantic Versioning](https://doc.rust-lang.org/
 
 - Curves:
   - added Curve: main interface
+  - added ConstCurve: curves with constant value
   - added Track, FrameIndex, TrackContent: concrete implementation
   - added CurveBuilder, TrackContentBuilder: convenience API
 - Animations:
@@ -11,3 +12,6 @@ This project adheres to [Cargo's Semantic Versioning](https://doc.rust-lang.org/
   - added AnimationPlayer, AnimationPlugin: worker
 - added optics-based "path" for locating attributes to be animated
 - added Reflect-based "path" for locating attributes to be animated
+- added Transform2d: full-fledged affine transformation
+  - use in place of normal local Transform in Bevy
+  - require Bevy fork with GlobalTransform being Affine3A